correção

+++ estudo/routes/categoria.js
@@ -1,6 +1,5 @@
 const express = require("express")
 const mongoose = require("mongoose")
-const { jsx } = require("react/jsx-runtime")
 const router = express.Router()
 require("../models/categoria")
 const cat = mongoose.model("categorias")
@@ -133,6 +132,9 @@ router.put("/:id", async (req, res)=>{
 })
 router.delete("/:id", async (req, res)=>{
     try{
+        if(!mongoose.isValidObjectId(req.params.id)){
+           return res.status(404).json({mensagem: "Coloque um ID válido"}) 
+        }
         const del = await cat.findOneAndDelete({_id:req.params.id})
         if(!del){
           return res.status(404).json({mensagem: "ID não encontrado"})
